requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
scikit-surprise==1.1.3
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.4.2
numpy==1.24.4
mongoengine